backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
python-dotenv==1.0.0
SQLAlchemy==2.0.23
celery==5.3.4
loguru==0.7.2
openai>=1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
PyJWT==2.8.0
pytest==7.4.3
pytest-asyncio==0.21.1

